Improve menu item alignment
