+++ packages/story-editor/src/components/carousel/gridview/gridViewButton.js
@@ -13,27 +13,29 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 /**
  * External dependencies
  */
 import { __ } from '@web-stories-wp/i18n';
-import { useCallback } from '@web-stories-wp/react';
-import { trackEvent } from '@web-stories-wp/tracking';
-import { Tooltip } from '@web-stories-wp/story-editor';
 import {
   Button,
   BUTTON_SIZES,
   BUTTON_TYPES,
   BUTTON_VARIANTS,
   Icons,
+  Modal,
   PLACEMENT,
 } from '@web-stories-wp/design-system';
+import { useCallback, useState } from '@web-stories-wp/react';
+import { trackEvent } from '@web-stories-wp/tracking';
 import styled from 'styled-components';
 
 /**
  * Internal dependencies
  */
-import useMetaBoxes from '../useMetaBoxes';
+import Tooltip from '../../tooltip';
+import GridView from './gridView';
 
 const Box = styled.div`
   width: 32px;
@@ -43,32 +45,26 @@ const Box = styled.div`
   align-items: center;
 `;
 
-const Item = styled.div``;
+function GridViewButton() {
+  const [isGridViewOpen, setIsGridViewOpen] = useState(false);
 
-function MenuItem() {
-  const { metaBoxesVisible, toggleMetaBoxesVisible, hasMetaBoxes } =
-    useMetaBoxes(({ state, actions }) => ({
-      hasMetaBoxes: state.hasMetaBoxes,
-      metaBoxesVisible: state.metaBoxesVisible,
-      toggleMetaBoxesVisible: actions.toggleMetaBoxesVisible,
-    }));
+  const toggleModal = useCallback(() => {
+    setIsGridViewOpen((prevIsOpen) => {
+      const newIsOpen = !prevIsOpen;
 
-  const handleMetaBoxesClick = useCallback(() => {
-    toggleMetaBoxesVisible();
-    trackEvent('meta_boxes_toggled', {
-      status: metaBoxesVisible ? 'visible' : 'hidden',
+      trackEvent('grid_view_toggled', {
+        status: newIsOpen ? 'open' : 'closed',
+      });
+
+      return newIsOpen;
     });
-  }, [metaBoxesVisible, toggleMetaBoxesVisible]);
-
-  if (!hasMetaBoxes) {
-    return null;
-  }
+  }, [setIsGridViewOpen]);
 
   return (
-    <Item>
+    <>
       <Box>
         <Tooltip
-          title={__('Third-Party Meta Boxes', 'web-stories')}
+          title={__('Grid View', 'web-stories')}
           placement={PLACEMENT.TOP}
           hasTail
         >
@@ -76,15 +72,30 @@ function MenuItem() {
             variant={BUTTON_VARIANTS.SQUARE}
             type={BUTTON_TYPES.TERTIARY}
             size={BUTTON_SIZES.SMALL}
-            onClick={handleMetaBoxesClick}
-            aria-label={__('Third-Party Meta Boxes', 'web-stories')}
+            onClick={toggleModal}
+            aria-label={__('Grid View', 'web-stories')}
           >
-            <Icons.LetterMOutline />
+            <Icons.Box4 />
           </Button>
         </Tooltip>
       </Box>
-    </Item>
+      <Modal
+        isOpen={isGridViewOpen}
+        onClose={toggleModal}
+        contentLabel={__('Grid View', 'web-stories')}
+        overlayStyles={{
+          alignItems: 'stretch',
+          backgroundColor: '#131516', // theme.colors.brand.gray[90]
+        }}
+        contentStyles={{
+          pointerEvents: 'none',
+          flex: 1,
+        }}
+      >
+        <GridView onClose={toggleModal} />
+      </Modal>
+    </>
   );
 }
 
-export default MenuItem;
+export default GridViewButton;
